agentzero-commerce/src/requirements.txt
fastapi
uvicorn[standard]
google-analytics-data
python-dotenv
sqlmodel
asyncpg
fastapi-utils
